+++ src/algo/63/smith_waterman_63.c
@@ -52,12 +52,12 @@ int64_t full_sw( sequence * dseq, sequence * qseq, int64_t * hearray ) {
         dbg_init_matrix_data_collection( BIT_WIDTH_64, dseq->len, qseq->len );
 #endif
 
-    for( uint64_t j = 0; j < dseq->len; j++ ) {
+    for( size_t j = 0; j < dseq->len; j++ ) {
         hep = hearray;
         f = 0;
         h = 0;
 
-        for( uint64_t i = 0; i < qseq->len; i++ ) {
+        for( size_t i = 0; i < qseq->len; i++ ) {
             n = *hep;
             e = *(hep + 1);
             h += SCORE_MATRIX_63( dseq->seq[j], qseq->seq[i] );
