+++ src/algo/aligner.h
@@ -14,10 +14,10 @@ void a_init_data( int search_type );
 
 void a_free_data();
 
-void a_free(p_alignment_list alist);
+void a_free( p_alignment_list alist );
 
-void a_set_alignment_pairs( int pair_count, elem_t * result_sequence_pairs );
+void a_set_alignment_pairs( size_t pair_count, elem_t * result_sequence_pairs );
 
-void * a_align(void * adp);
+void * a_align( void * adp );
 
 #endif /* ALIGNER_H_ */
