+++ src/util/linked_list.h
@@ -8,6 +8,8 @@
 #ifndef SRC_UTIL_LINKED_LIST_H_
 #define SRC_UTIL_LINKED_LIST_H_
 
+#include <stddef.h>
+
 struct node_t {
     void * data;
 
@@ -26,7 +28,7 @@ void ll_push( p_node * current, void * data );
 
 void * ll_pop( p_node * current );
 
-unsigned long ll_size( p_node node );
+size_t ll_size( p_node node );
 
 void ll_print( p_node * node );
